small changes in file structure, small fixes
