+++ Arduino/src/State/States.h
@@ -1,0 +1,6 @@
+#ifndef __STATES__
+#define __STATES__
+
+enum StateName { Idle, Welcome, CarExited };
+
+#endif
